test_approx.c: fix printf format

+++ test_approx.c
@@ -124,14 +124,14 @@ void result_data_print(struct result_data *result, int n)
 {
     int i;
 
-    fprintf(stderr, "%d", result->ticks[0]/n);
+    fprintf(stderr, "%g", (double)result->ticks[0]/n);
     for (i = 1; i < nr_methods; ++i)
-	fprintf(stderr, ", %d", result->ticks[i]/n);
+	fprintf(stderr, ", %g", (double)result->ticks[i]/n);
     fprintf(stderr, "\n");
 
-    fprintf(stderr, "%d", result->size[0]/n);
+    fprintf(stderr, "%zd", result->size[0]/n);
     for (i = 1; i < nr_methods; ++i)
-	fprintf(stderr, ", %d", result->size[i]/n);
+	fprintf(stderr, ", %zd", result->size[i]/n);
     fprintf(stderr, "\n");
 
     fprintf(stderr, "%g\n", VALUE_TO_DOUBLE(result->n));
